chore: add required fields to the data model

+++ app/composeApp/src/commonMain/kotlin/com/altmanifest/secondtake/domain/Rating.kt
@@ -3,8 +3,8 @@ package com.altmanifest.secondtake.domain
 import kotlin.math.abs
 import kotlin.time.Duration
 
-data class Rating(val value: Int, val age: Duration) {
-    fun pointDifference(other: Rating): Int = abs(other.value - this.value)
+data class Rating(val value: Double, val age: Duration) {
+    fun pointDifference(other: Rating): Double = abs(other.value - this.value)
     fun comparableWith(other: Rating, tolerance: Int): Boolean = pointDifference(other) <= tolerance
     fun notComparableWith(other: Rating, tolerance: Int): Boolean = !comparableWith(other, tolerance)
 }
